% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekostiging.R
\docType{data}
\name{BEKOSTIGINGSTATUS_CODES}
\alias{BEKOSTIGINGSTATUS_CODES}
\title{Redencodes bekostigingstatus (VLPBEK/DEFBEK)}
\format{
Een tibble met kolommen \code{code}, \code{omschrijving} en \code{herstelbaar}
}
\usage{
BEKOSTIGINGSTATUS_CODES
}
\description{
Decodeertabel voor de kolom \code{CodeBekostigingstatus} uit het analysebestand
voorlopige/definitieve bekostiging (DUO PvE HO-instelling, bijlage 8,
par. 17.7.3/19.7.5). Een inschrijving of graad kan meerdere codes
tegelijk hebben (komma-gescheiden in het brondata, bijv. \code{"na,ti"}).
}
\details{
\code{herstelbaar} geeft aan of de reden voortkomt uit een registratiefout van
de instelling die te herstellen is door tijdig opnieuw aan te leveren
(\code{ti}/\code{tg}, waarde \code{TRUE}), een structurele/wettelijke reden is die niet
te herstellen is (waarde \code{FALSE}), of helemaal geen "niet bekostigd"-reden
betreft maar juist aangeeft dat er (deels) bekostigd wordt (waarde \code{NA}).
}
\keyword{datasets}
