% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staat1cho-package.R
\docType{package}
\name{staat1cho-package}
\alias{staat1cho}
\alias{staat1cho-package}
\title{staat1cho: Studie-indicatoren op basis van 1CHO-data}
\description{
Berekent instroom-, rendement-, uitval- en studiewisselindicatoren
vanuit de 1CHO-aanlevering van DUO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cedanl/staat-van-onderwijsinstelling}
  \item Report bugs at \url{https://github.com/cedanl/staat-van-onderwijsinstelling/issues}
}

}
\author{
\strong{Maintainer}: Aslam

Authors:
\itemize{
  \item Veerle van Son
  \item Damiëtte Bakx-van den Brink
}

}
\keyword{internal}
