% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendement.R
\name{bereken_rendement}
\alias{bereken_rendement}
\title{Bereken rendementsindicatoren per cohort}
\usage{
bereken_rendement(cohorten_instroom, diploma_behaald)
}
\arguments{
\item{cohorten_instroom}{Tibble zoals gemaakt door \code{\link[=maak_instroom_cohort]{maak_instroom_cohort()}}}

\item{diploma_behaald}{Tibble zoals gemaakt door \code{\link[=maak_diploma_behaald]{maak_diploma_behaald()}}}
}
\value{
Een tibble met kolommen \code{persoonsgebonden_nummer},
\code{eerstejaar_instelling}, \code{jaar_eerste_diploma},
\code{verblijfsjaar_eerste_diploma}, \code{diploma}, \code{rendement_xjaar},
en factorkolommen \code{rendement_3jr}, \code{rendement_5jr}, \code{rendement_8jr}
}
\description{
Koppelt diplomagegevens aan het instroomcohort en berekent of een student
binnen 3, 5 of 8 jaar een diploma heeft behaald.
}
