% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiewissel.R
\name{bereken_studiewissel}
\alias{bereken_studiewissel}
\title{Bereken studiewisselindicatoren per cohort}
\usage{
bereken_studiewissel(
  basisbestand,
  cohorten_instroom,
  diploma_behaald,
  uitval_indicatoren
)
}
\arguments{
\item{basisbestand}{Tibble zoals gemaakt door \code{\link[=maak_basisbestand]{maak_basisbestand()}}}

\item{cohorten_instroom}{Tibble zoals gemaakt door \code{\link[=maak_instroom_cohort]{maak_instroom_cohort()}}}

\item{diploma_behaald}{Tibble zoals gemaakt door \code{\link[=maak_diploma_behaald]{maak_diploma_behaald()}}}

\item{uitval_indicatoren}{Tibble zoals gemaakt door \code{\link[=bereken_uitval]{bereken_uitval()}}}
}
\value{
Een tibble met kolommen \code{persoonsgebonden_nummer},
\code{studiewissel_1jr}, \code{studiewissel_3jr} (factoren) en aanvullende
switch-kolommen met de opleiding, opleidingsvorm, niveau en sector na
de wissel
}
\description{
Bepaalt per student of zij binnen 1 of 3 jaar van opleiding zijn gewisseld.
Studenten die al zijn uitgevallen of gediplomeerd in de meetperiode worden
buiten beschouwing gelaten.
}
