% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instroom.R
\name{maak_basisbestand}
\alias{maak_basisbestand}
\title{Lees het 1CHO-bestand in en voeg label-kolommen toe}
\usage{
maak_basisbestand(pad_invoer)
}
\arguments{
\item{pad_invoer}{Pad naar het semicolongescheiden CSV-bestand (UTF-8)}
}
\value{
Een tibble met alle 1CHO-regels plus extra \verb{_label}-kolommen die de
originele categorische waarden bewaren voor gebruik in rapportages
}
\description{
Lees het 1CHO-bestand in en voeg label-kolommen toe
}
