% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendement.R
\name{maak_diploma_behaald}
\alias{maak_diploma_behaald}
\title{Maak een bestand met het vroegst behaalde diploma per student}
\usage{
maak_diploma_behaald(basisbestand)
}
\arguments{
\item{basisbestand}{Tibble zoals gemaakt door \code{\link[=maak_basisbestand]{maak_basisbestand()}}}
}
\value{
Een tibble met één rij per student met kolommen
\code{persoonsgebonden_nummer}, \code{jaar_eerste_diploma},
\code{verblijfsjaar_eerste_diploma} en \code{diploma}. Gooit een fout bij
dubbele persoonsgebonden nummers.
}
\description{
Filtert het basisbestand op diplomasoorten die gelden als afgeronde opleiding
en behoudt per student alleen het eerste diploma op basis van diplomajaar.
}
