% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{start_dashboard}
\alias{start_dashboard}
\title{Start het Staat van Onderwijsinstelling dashboard}
\usage{
start_dashboard()
}
\description{
Opent de interactieve Shiny app waarmee je een 1CHO CSV-bestand kunt
uploaden en studie-indicatoren kunt verkennen.
}
