% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekostiging.R
\name{verrijk_met_bekostiging}
\alias{verrijk_met_bekostiging}
\title{Verrijk indicatoren met bekostigingsinformatie uit een VLPBEK-bestand}
\usage{
verrijk_met_bekostiging(indicatoren, bekostiging)
}
\arguments{
\item{indicatoren}{Tibble zoals gemaakt door \code{\link[=combineer_indicatoren]{combineer_indicatoren()}},
met kolommen \code{persoonsgebonden_nummer} en \code{opleidingscode}}

\item{bekostiging}{Tibble zoals gemaakt door \code{\link[=lees_bekostiging]{lees_bekostiging()}}}
}
\value{
De indicatoren-tibble uitgebreid met \code{indicatie_bekostigd}
(\code{TRUE}/\code{FALSE}/\code{NA}), \code{indicatie_hoofdinschrijving} (\code{TRUE}/\code{FALSE}/\code{NA}),
\code{reden_niet_bekostigd} (leesbare toelichting(en), \code{NA} als wel bekostigd)
en \code{indicatie_herstelbaar} (\code{TRUE}/\code{FALSE}/\code{NA}, zie
\link{BEKOSTIGINGSTATUS_CODES}). Heeft een student meerdere BRD-regels met
verschillende redenen voor dezelfde opleiding, dan worden de unieke
redenen samengevoegd.
}
\description{
Koppelt bekostigingsstatus per student per opleiding aan het
indicatorenbestand via \code{persoonsgebonden_nummer} en \code{opleidingscode}.
Wanneer een student meerdere BRD-regels heeft voor dezelfde opleiding
geldt: als ten minste een regel de status \code{"bekostigd"} heeft, is de
student bekostigd voor die opleiding.
}
\details{
Studenten zonder overeenkomst in het VLPBEK-bestand krijgen \code{NA} voor
\code{indicatie_bekostigd} en \code{indicatie_hoofdinschrijving}.
}
\examples{
indicatoren <- tibble::tibble(
  persoonsgebonden_nummer = c("1", "2", "3"),
  opleidingscode          = c("31001", "31002", "31003"),
  inschrijvingsjaar       = 2023L
)
bekostiging <- tibble::tibble(
  persoonsgebonden_nummer    = c("1", "2"),
  opleidingscode             = c("31001", "31002"),
  inschrijvingsjaar          = c(2023L, 2023L),
  indicatie_hoofdinschrijving = c(TRUE, TRUE),
  opleidingsvorm             = c("voltijd", "voltijd"),
  sector                     = c("gezondheidszorg", "economie"),
  bekostigingsstatus         = c("bekostigd", "niet bekostigd"),
  code_bekostigingstatus     = c(NA, "nf"),
  reden_niet_bekostigd       = c(NA, "Niet bekostigd i.v.m. overschrijden maximaal aantal bekostigde inschrijvingen, rekening houdend met eerder behaalde graden."),
  indicatie_herstelbaar      = c(NA, FALSE)
)
verrijk_met_bekostiging(indicatoren, bekostiging)
}
